Final refactoring and cleanups
