+++ genienlp/models/transformer_lstm.py
@@ -199,11 +199,12 @@ class TransformerLSTM(GenieModel):
                                      generation_dict={'max_output_length': max_output_length},
                                      encoder_output=encoder_output,
                                      output_scores=False,
-                                     output_attentions=True,
+                                     output_attentions=False,
                                      output_hidden_states=False,
                                      return_dict_in_generate=True
                                      )
         output_ids = generated.sequences
-        output_ids = torch.cat((output_ids[:, 0:1], output_ids[:, 1:].cpu().apply_(self.decoder.map_to_full).to(batch.context.value.device)), dim=1) # map everything to full vocabulary except BOS which already is in full vocabulary
+        mapped_output_ids = torch.cat((output_ids[:, 0:1], output_ids[:, 1:].cpu().apply_(self.decoder.map_to_full).to(batch.context.value.device)), dim=1) # map everything to full vocabulary except BOS which already is in full vocabulary
+        generated.sequences = mapped_output_ids
         
-        return output_ids
+        return generated
