+++ genienlp/util.py
@@ -663,7 +663,8 @@ def load_config_json(args):
                     'ned_features', 'ned_features_size', 'ned_features_default_val',
                     'bootleg_output_dir', 'bootleg_model', 'bootleg_batch_size',
                     'bootleg_kg_encoder_layer', 'bootleg_dataset_threads', 'bootleg_dataloader_threads', 'bootleg_extract_num_workers',
-                    'bootleg_dump_mode', 'bootleg_prob_threshold', 'bootleg_post_process_types'
+                    'bootleg_dump_mode', 'bootleg_prob_threshold', 'bootleg_post_process_types',
+                    'att_pooling', 'plot_heatmaps', 'replace_qp', 'force_replace_qp'
                     ]
 
         # train and predict scripts have these arguments in common. We use the values from train only if they are not provided in predict
@@ -680,7 +681,8 @@ def load_config_json(args):
                 setattr(args, r, config[r])
             # These are for backward compatibility with models that were trained before we added these arguments
             elif r in ('do_ned', 'use_encoder_loss',
-                       'almond_has_multiple_programs', 'almond_lang_as_question', 'preprocess_special_tokens', 'almond_thingtalk_version'
+                       'almond_has_multiple_programs', 'almond_lang_as_question', 'preprocess_special_tokens', 'almond_thingtalk_version',
+                       'plot_heatmaps', 'replace_qp', 'force_replace_qp'
                        ):
                 setattr(args, r, False)
             elif r in ('num_db_types', 'db_unk_id', 'num_workers'):
@@ -697,6 +699,8 @@ def load_config_json(args):
                 setattr(args, r, 'no')
             elif r == 'database_type':
                 setattr(args, r, 'json')
+            elif r == 'att_pooling':
+                setattr(args, r, 'max')
             elif r == 'min_entity_len':
                 setattr(args, r, 2)
             elif r == 'max_entity_len':
