+++ tests/dataset/translation/en-de/dev_marian_aligned.tsv
@@ -3,4 +3,4 @@ show me nearby hotels with both a " catalan " and " sauna "	zeigen Sie mir in de
 find people graduate of Stanford.	Leute finden, die Stanford graduieren.
 what is the highest rated hotel ?	was ist das am höchsten bewertete Hotel ?
 find hotels with 2 star ratings .	Hotels mit 2 Sterne Bewertungen finden.
-what is the rating of " rosedon " in " glenorchy " ?	Wie hoch ist die Bewertung von " rosedon " " in " glenorchy " " ?
+what is the rating of " rosedon " in " glenorchy " ?	Wie hoch ist die Bewertung von " rosedon " in " glenorchy " ?
